add syntax highlight language selection dropdown
